Add title logic in header component
